Fix CustomServerPaintingsState & Painting entity
